real-time-fraud-detection/requirements.txt
fastapi
uvicorn
kafka-python
scikit-learn
pandas
numpy
joblib
mlflow
psutil
feast==0.49.0
redis
httpx
requests